src/rag_lambda/requirements.txt
# Lambda dependencies for RAG application
langchain>=0.2.0
langgraph>=0.2.0
langchain-aws>=1.0.0
langchain-postgres>=0.0.10
langchain-community>=0.2.0
psycopg[binary]>=3.1.0
boto3>=1.28.0
pydantic>=2.0.0
pyyaml>=6.0

